Route grouped experts to Muon + fix AMoE autocast dtype scatter
- optim.py: split_params rule p.ndim==2 -> p.ndim>=2 so the 3D grouped-expert
  tensors (expert1/expert2) train under Muon (batched Newton-Schulz over the last
  2 dims), not AdamW. Muon group 6.3M -> 56.7M.
- model.py: in AMoE.forward, cast the sparse MoE outputs to the destination dtype
  on scatter (new_flat/cert_flat) so fp16/bf16 autocast training no longer crashes
  on a Float/Half index_put mismatch. This is the ONLY model.py change.
- Verified on H100: check_model.py 7/7 PASS (incl. HF Trainer fp16 step); Muon
  now orthogonalizes the 3D expert matrices.

+++ model.py
@@ -194,8 +194,10 @@ class AMoE(nn.Module):
                 new_active, cert_active = self._moe_call(active_tokens)
                 
                 # 4. 연산된 결과를 원래 배치 위치로 복원 (Scatter)
-                new_flat[active_mask] = new_active
-                cert_flat[active_mask] = cert_active
+                # autocast(fp16/bf16)에서 MoE 출력 dtype이 new_flat/cert_flat(fp32)과
+                # 다를 수 있어 dst dtype에 맞춰 캐스팅 (누적은 fp32로 유지)
+                new_flat[active_mask] = new_active.to(new_flat.dtype)
+                cert_flat[active_mask] = cert_active.to(cert_flat.dtype)
 
             # 다시 3차원으로 복구
             new_state = new_flat.view(B, N, D)
